+++ src/modules/midi/MidiStep.cpp
@@ -92,6 +92,15 @@ struct MidiStepModule : Module {
 		Module::onReset(e);
 	}
 
+	void processBypass(const ProcessArgs& args) override {
+		midi::Message msg;
+		// Drain the queue while bypassed
+		while (midiInput.tryPop(&msg, args.frame)) {
+			(void)0;
+		}
+		Module::processBypass(args);
+	}
+
 	void process(const ProcessArgs& args) override {
 		midi::Message msg;
 		while (midiInput.tryPop(&msg, args.frame)) {